feat: A2A Protocol, WASM Plugin Architecture, Governance Matrix
- A2A Protocol: Google A2A-compatible inter-agent communication with Agent Card
  discovery, task lifecycle, and ACP transport bridge
- Plugin Architecture: WASM-based plugin system with wasmtime sandbox,
  TOML/JSON manifests, and host access control
- Governance Matrix: per-agent resource access control with Allow/Deny/Ask
  matrix, ToolGuard (AllowList/BlockList), and FileGuard (restricted paths)
- Update README and ARCHITECTURE docs
- Add no-worktree rule to AGENTS.md
- Remove leftover git worktrees

+++ crates/core/src/tools/time_tool.rs
@@ -37,7 +37,7 @@ impl Tool for TimeTool {
 
         Ok(json!({
             "unix_timestamp_secs": unix_secs,
-            "unix_timestamp_ms": unix_secs as u64 * 1000 + millis as u64,
+            "unix_timestamp_ms": unix_secs * 1000 + millis as u64,
             "date_iso": iso_date(unix_secs),
             "timezone": "UTC",
         }))
